requirements.txt
flask
flask-sqlalchemy
flask-login
email-validator
Pillow